[agent_builder] - feature: enhance AccessSection with Slack settings context
 - Integrate `useDataSourceViewsContext` to conditionally display Slack preferences in `AccessSection`
 - Update `SlackSettingsSheet` to accept `slackDataSource` as a prop for improved context handling
 - Code maintenance in `AccessSection` to use destructured `scope` field for readability and consistency
